requirements.txt
PyYAML>=6.0
pydantic>=2.6,<3.0
psutil>=5.9,<6.0
Jinja2>=3.1,<4.0
